+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/clients/anthropic/AnthropicClientConfiguration.kt
@@ -38,7 +38,7 @@ class AnthropicClientConfiguration : LLMClientConfiguration(
     }
 
     override fun getClientIcon(): Icon {
-        return Icons.ANTHROPIC
+        return Icons.ANTHROPIC.getThemeBasedIcon()
     }
 
     override fun getSharedState(): LLMClientSharedState {