+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/clients/ollama/OllamaClientConfiguration.kt
@@ -30,7 +30,7 @@ class OllamaClientConfiguration : LLMClientConfiguration(
     }
 
     override fun getClientIcon(): Icon {
-        return Icons.OLLAMA
+        return Icons.OLLAMA.getThemeBasedIcon()
     }
 
     override fun getSharedState(): LLMClientSharedState {